Restructure the browser extension source tree
The previous structure was organized by browser on the assumption that
there would be a significant amount of browser-specific code/manifests
and some shared files (in `src/common`). In practice every browser that supports
extensions (Chrome, variants of Chrome, Firefox) all use Chrome's basic
APIs.

Therefore the src/ directory has been re-arranged to group files by
which _part_ of the extension they are for, with general
manifest/settings etc. files at the top level.

 - The source files for the extension's background page have been moved
   from src/common to src/background and the entry point has been
   renamed to index.js
 - The files in src/chrome have been moved to src/ as they are used by
   all browsers
 - The resources for the extension options page are now grouped into an
   options directory
 - app.html.mustache has been renamed to sidebar-app.html.mustache for
   clarity
 - destroy.js has been renamed to unload-client.js for clarity
