+++ src/background/browser-action.js
