+++ src/background/browser-name.js
