+++ src/background/detect-content-type.js
