+++ src/background/direct-link-query.js
