+++ src/background/errors.js
