+++ src/background/help-page.js
