+++ src/background/hypothesis-chrome-extension.js
