+++ src/background/index.js
