+++ src/background/install.js
