+++ src/background/polyfills.js
