+++ src/background/settings.js
