+++ src/background/tab-state.js
