+++ src/background/tab-store.js
