+++ src/background/uri-info.js
