+++ src/background/util.js
