+++ src/images/browser-icon-active.png
