+++ src/images/browser-icon-active@2x.png
