+++ src/images/browser-icon-inactive@2x.png
