+++ src/manifest.json.mustache
@@ -40,7 +40,7 @@
   {{/browserIsChrome}}
 
   "options_ui": {
-    "page": "lib/options.html",
+    "page": "options/index.html",
     "chrome_style": true
   },
 
@@ -80,7 +80,6 @@
 
   "web_accessible_resources": [
     "client/*",
-    "lib/*",
     "help/*",
     "pdfjs/*",
     "pdfjs/web/viewer.html"
