+++ src/settings-data.js.mustache
