+++ src/sidebar-app.html.mustache
